Guerrilla defensive programming
I swear I saw this thing crash once or twice while Narrator was starting
up and I was in the middle of typing out a filter, but never ever could
I get it to fail under a debugger. Maybe it was nothing and I was imagining
things.

In any case, the only race I can think of is trying to access slots from
a ThumbnailUIAProvider while the slots array is being rebuilt. Hence, added
an atomic int to signal this. Since it's not a mutex, it's still racy, but
the odds of hitting are a bit slimmer, I think?

I don't really want to add a full blown mutex until I get a clear explanation
of what sort of threads UIA injects into my single threaded program, b/c this
is a bit ridiculous to try to reverse engineer (especially if it's not contractual)

Anyway, should be slightly --but not entirely-- more stable under screen reader
AT now.